Cleaned up feature files
