+++ src/test/java/com/mysecondcucumberproject/pageObject/AutomationPracticeHomePage.java
@@ -5,14 +5,16 @@ import java.util.List;
 
 import org.openqa.selenium.By;
 import org.openqa.selenium.NoSuchElementException;
-import org.openqa.selenium.NoSuchFrameException;
 import org.openqa.selenium.Point;
 import org.openqa.selenium.WebDriver;
 import org.openqa.selenium.WebElement;
 import org.openqa.selenium.interactions.Actions;
 import org.openqa.selenium.interactions.MoveTargetOutOfBoundsException;
 import org.openqa.selenium.support.FindBy;
+import org.openqa.selenium.support.ui.Select;
+import org.openqa.selenium.support.ui.UnexpectedTagNameException;
 
+import com.mysecondcucumberproject.factory.BaseUtilities;
 import com.mysecondcucumberproject.utilities.TestConstants;
 
 public class AutomationPracticeHomePage extends BasePage {
@@ -69,8 +71,6 @@ public class AutomationPracticeHomePage extends BasePage {
 	// #region Paginated table
 	@FindBy(xpath = "//*[@id=\"productTable\"]")
 	WebElement paginatedTable;
-	@FindBy(xpath = "//*[@id=\"productTable\"]/tbody")
-	WebElement paginatedTableBody;
 	@FindBy(xpath = "//*[@id=\"pagination\"]")
 	WebElement paginationButtonsField;
 	@FindBy(xpath = "//*[@id=\"pagination\"]//a")
@@ -90,7 +90,8 @@ public class AutomationPracticeHomePage extends BasePage {
 				return columnIndex;
 		}
 		System.out.println("Couldn't find any header of column with the searchterm: " + searchTerm);
-		return columnIndex = 0;
+		columnIndex = 0;
+		return columnIndex;
 	}
 
 	// #region Tabs
@@ -162,14 +163,31 @@ public class AutomationPracticeHomePage extends BasePage {
 	// button"
 	@FindBy(xpath = "//h2[contains(@class, \"title\") and contains(text(), \"Form\")]/..//*[contains(@class, \"rectangular-button\")]")
 	List<WebElement> inputSubmitButtons;
-
 	// #endregion
 
-	// #region Resizable
-	@FindBy(xpath = "//div[h2[@class='title' and text()='Resizable']]")
-	WebElement resizeableContainer;
-	@FindBy(xpath = "//*[@id=\"resizable\"]")
-	WebElement resizableElement;
+	// #region date of birth
+	@FindBy(xpath = "//*[@id=\"q4\"]")
+	WebElement dOBContainer;
+	@FindBy(xpath = "//*[@id='RESULT_TextField-2']")
+	WebElement dOBField;
+	@FindBy(xpath = "//span=[@class='icon_calendar']")
+	WebElement calendarButton;
+	@FindBy(xpath = "//*[@id=\"RESULT_RadioButton-3\"]")
+	WebElement workDropDown;
+	@FindBy(xpath = "//*[@id='ui-datepicker-div']")
+	WebElement dobDropdown;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/div/a[1]/span")
+	WebElement dobPastMonthButton;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/div/a[2]")
+	WebElement dobComingMonthButton;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/div/div/select")
+	WebElement dobYearDropDown;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/div/div/span")
+	WebElement dobCurrentMonthText;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/table/tbody")
+	WebElement dobTableBody;
+	@FindBy(xpath = "//*[@id=\"ui-datepicker-div\"]/table/tbody/td")
+	List<WebElement> dateBoxes;
 	// #endregion
 
 	// Action methods
@@ -275,6 +293,11 @@ public class AutomationPracticeHomePage extends BasePage {
 			case TestConstants.FIELD_2_ID -> {
 				field2.click();
 				field2.sendKeys(input);
+				return true;
+			}
+			case TestConstants.FRAMEDOBFIELD_ID -> {
+				dOBField.clear();
+				dOBField.sendKeys(input);
 				return true;
 			}
 			default -> {
@@ -392,6 +415,36 @@ public class AutomationPracticeHomePage extends BasePage {
 			case TestConstants.PRACTICEFORMCONTAINTER_ID -> {
 				return practiceFormFrame.isDisplayed();
 			}
+			case TestConstants.DOBCONTAINER_ID -> {
+				return (dOBContainer.isDisplayed() && dOBContainer.isEnabled());
+			}
+			case TestConstants.FRAMEDOBFIELD_ID -> {
+				return (dOBField.isDisplayed() && dOBField.isEnabled());
+			}
+			case TestConstants.CALENDARBUTTON_ID -> {
+				return (calendarButton.isDisplayed() && calendarButton.isEnabled());
+			}
+			case TestConstants.FRAMEWORKDROPDOWN_ID -> {
+				return (workDropDown.isDisplayed() && workDropDown.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWN_ID -> {
+				return (dobDropdown.isDisplayed() && dobDropdown.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWNPASTMONTH_ID -> {
+				return (dobPastMonthButton.isDisplayed() && dobPastMonthButton.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWNCOMINGMONTH_ID -> {
+				return (dobComingMonthButton.isDisplayed() && dobComingMonthButton.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWNCURRENTMONTH_ID -> {
+				return (dobComingMonthButton.isDisplayed() && dobComingMonthButton.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWNYEARDROPDOWN_ID -> {
+				return (dobYearDropDown.isDisplayed() && dobYearDropDown.isEnabled());
+			}
+			case TestConstants.DOBDROPDOWNTABLEBODY_ID -> {
+				return (dobTableBody.isDisplayed() && dobTableBody.isEnabled());
+			}
 			default -> {
 				System.out.println("Couln't find a webelement using: " + fieldID);
 				return false;
@@ -490,6 +543,39 @@ public class AutomationPracticeHomePage extends BasePage {
 			case TestConstants.PRACTICEFORMCONTAINTER_ID -> {
 				return practiceFormFrame;
 			}
+			case TestConstants.FRAMENAMEFIELD_ID -> {
+				return nameField;
+			}
+			case TestConstants.DOBCONTAINER_ID -> {
+				return dOBContainer;
+			}
+			case TestConstants.FRAMEDOBFIELD_ID -> {
+				return dOBField;
+			}
+			case TestConstants.CALENDARBUTTON_ID -> {
+				return calendarButton;
+			}
+			case TestConstants.DOBDROPDOWN_ID -> {
+				return dobDropdown;
+			}
+			case TestConstants.DOBDROPDOWNPASTMONTH_ID -> {
+				return dobPastMonthButton;
+			}
+			case TestConstants.DOBDROPDOWNCOMINGMONTH_ID -> {
+				return dobComingMonthButton;
+			}
+			case TestConstants.DOBDROPDOWNCURRENTMONTH_ID -> {
+				return dobComingMonthButton;
+			}
+			case TestConstants.DOBDROPDOWNYEARDROPDOWN_ID -> {
+				return dobYearDropDown;
+			}
+			case TestConstants.DOBDROPDOWNTABLEBODY_ID -> {
+				return dobTableBody;
+			}
+			case TestConstants.FRAMEWORKDROPDOWN_ID -> {
+				return workDropDown;
+			}
 			default -> {
 				System.out.println("Couln't find a webelement using: " + fieldID);
 				return null;
@@ -529,13 +615,13 @@ public class AutomationPracticeHomePage extends BasePage {
 	public String getElementText(String elementID) {
 		switch (elementID.toLowerCase().trim()) {
 			case TestConstants.NAMEFIELD_ID -> {
-				return nameField.getAttribute("value");
+				return nameField.getDomProperty("value");
 			}
 			case TestConstants.EMAILFIELD_ID -> {
-				return emailField.getAttribute("value");
+				return emailField.getDomProperty("value");
 			}
 			case TestConstants.PHONEFIELD_ID -> {
-				return phoneField.getAttribute("value");
+				return phoneField.getDomProperty("value");
 			}
 			case TestConstants.PAGINATEDTABLETOPLEFTMOSTCELL_ID -> {
 				return topLeftmostProductTableCell.getText();
@@ -546,6 +632,9 @@ public class AutomationPracticeHomePage extends BasePage {
 			case TestConstants.FIELD_2_ID -> {
 				return field2.getText();
 			}
+			case TestConstants.FRAMEDOBFIELD_ID -> {
+				return dOBField.getDomAttribute("value");
+			}
 			default -> {
 				System.out.println(this + "couldn't find an element to get the value of with the ID of: " + elementID);
 				return null;
@@ -554,7 +643,7 @@ public class AutomationPracticeHomePage extends BasePage {
 	}
 
 	public boolean isPaginationButtonSelected(int index) {
-		return productTablePageButtons.get(index - 1).getAttribute("class").contains("active") == true ? true : false;
+		return productTablePageButtons.get(index - 1).getDomAttribute("class").contains("active") == true;
 	}
 
 	public int getProductPageAmount() {
@@ -584,7 +673,7 @@ public class AutomationPracticeHomePage extends BasePage {
 			arr[i] = weekdayCheckboxes.get(i);
 		}
 
-		return (arr[index].isSelected()) == true ? true : false;
+		return (arr[index].isSelected()) == true;
 	}
 
 	public String getWeekdayCheckboxValue(int index) {
@@ -596,7 +685,7 @@ public class AutomationPracticeHomePage extends BasePage {
 			arr[i] = weekdayCheckboxes.get(i);
 		}
 
-		return arr[index].getAttribute("value");
+		return arr[index].getDomAttribute("value");
 	}
 
 	public boolean tryClickButton(String buttonID) {
@@ -698,22 +787,122 @@ public class AutomationPracticeHomePage extends BasePage {
 		}
 	}
 
-	public PracticeFormPage goToPracticeForm() {
+	public boolean trySelectInDropdown(String userSelection, String elementID) {
 
-		// TODO: Might want to remove this whole method, seeing as we aren't using
-		// iFrame anymore.
+		if (!canFindWebelement(elementID))
+			return false;
+
+		Select dropDown;
+
 		try {
-			driver.switchTo().frame(practiceFormFrame);
-
-			// TODO: What if it is another error then? Should I always have a catch for
-			// general errors?
-		} catch (NoSuchFrameException e) {
+			dropDown = new Select(getWebelement(elementID));
+		} catch (UnexpectedTagNameException e) {
+			System.out.println("Couldn't create a select object from the web element");
 			System.out.println(e.getMessage());
-			return null;
+			return false;
 		}
 
-		PracticeFormPage temp = new PracticeFormPage(driver);
+		try {
+			dropDown.selectByVisibleText(userSelection);
+		} catch (NoSuchElementException e) {
+			System.out.println("Couldn't find an element using this value.");
+			System.out.println(e.getMessage());
+			return false;
+		}
 
-		return temp;
+		return true;
 	}
+
+	public boolean trySelectInDropdown(String elementID, String year, int month, String date) {
+		if (!canFindWebelement(elementID))
+			return false;
+
+		Select dropDown;
+
+		try {
+			dropDown = new Select(getWebelement(elementID));
+		} catch (UnexpectedTagNameException e) {
+			System.out.println("Couldn't create a select object from the web element");
+			System.out.println(e.getMessage());
+			return false;
+		}
+
+		if (!trySelectInDropdown(year, TestConstants.DOBDROPDOWNYEARDROPDOWN_ID)) {
+			return false;
+		}
+
+		// The takeScreenShot() below is just during testing/working
+		takeScreenShot(TestConstants.DOBDROPDOWN_ID);
+
+		if (!trySetDOBMonth(dropDown, month)) {
+			return false;
+		}
+
+		// The takeScreenShot() below is just during testing/working
+		takeScreenShot(TestConstants.DOBDROPDOWN_ID);
+
+		if (!trySetDOBDay(dropDown, date)) {
+			return false;
+		}
+
+		// The takeScreenShot() below is just during testing/working
+		takeScreenShot(TestConstants.DOBDROPDOWN_ID);
+
+		return true;
+	}
+
+	private boolean trySetDOBDay(Select dropDown, String date) {
+
+		/*
+		 * See if it can find the needed elements.
+		 * Will otherwise have a list of data cells
+		 * go through them one by one, find it they have the right value.
+		 * If they do, do something to set it?
+		 */
+
+		/*
+		 * Rethink if this works, considering that these are a whole list of elements,
+		 * Also consider if the path for the body is even needed then?
+		 * Also consider if I can use a getter here, or it that doesn't work then?
+		 * Also consider, might not need dropDown here then I suppose?
+		 * Alse REMEMEBER that you want to go into each box child of type <a> to get the
+		 * inner text to get the date.
+		 */
+		if (!canFindWebelement(TestConstants.DOBDATEBOXES)) {
+			return false;
+		}
+
+		for (WebElement dateBox : dateBoxes) {
+			if (dateBox.findElement(By.xpath("/<a>")).getText().equals(date)) {
+				dateBox.click();
+				return true;
+			}
+		}
+
+		return false;
+	}
+
+	private boolean trySetDOBMonth(Select dropDown, int monthToSet) {
+		/*
+		 * SO, this one will use the main dropdown, wait, does it need that?
+		 * It will need the text field for months,
+		 * 
+		 * gets the month, as text, uses that to get the value of that month
+		 * compares to month
+		 */
+		while (BaseUtilities.getIndexOfMonth(dobCurrentMonthText.getText()) != monthToSet) {
+			if (BaseUtilities.getIndexOfMonth(dobCurrentMonthText.getText()) > monthToSet) {
+				// Decreases the current month by clicking on the button for past month
+				dobPastMonthButton.click();
+			} else if (BaseUtilities.getIndexOfMonth(dobCurrentMonthText.getText()) < monthToSet) {
+				// Increases the current month by clicking on the button for coming month
+				dobComingMonthButton.click();
+			}
+
+			System.out.println("The current month is(and we are working on changing that): " + dobCurrentMonthText);
+		}
+		// Will return true when the month has the correct value.
+		return true;
+	}
+
 }
