+++ src/test/java/com/mysecondcucumberproject/stepdefinitions/InteractingWithOtherElements.java
@@ -1,11 +1,8 @@
 package com.mysecondcucumberproject.stepdefinitions;
 
-import java.util.List;
-
 import org.junit.Assert;
 import org.openqa.selenium.Point;
 import org.openqa.selenium.WebDriver;
-import org.openqa.selenium.WebElement;
 
 import com.mysecondcucumberproject.factory.BaseUtilities;
 import com.mysecondcucumberproject.pageObject.AutomationPracticeHomePage;
@@ -331,98 +328,19 @@ public class InteractingWithOtherElements {
 
 	}
 
-	// TODO: Pretty sure this method can be removed now, as we are no longer messing
-	// about with iframes.
-	@Given("the user switches to {string}")
-	public void the_user_switches_to(String string) {
-
-		practiceFormPage = aPHomePage.goToPracticeForm();
-
-		try {
-			Assert.assertNotNull("Recieved a null page object when it attempted to go to practice form.",
-					practiceFormPage);
-		} catch (AssertionError e) {
-			System.out.println(e.getMessage());
-		}
-	}
-
-	@Given("the user enters {string} into each {string} field in the form")
-	public void the_user_enters_x_into_each_field_in_the_form(String userInput, String fieldID) {
-
-		/*
-		 * Get list from aPHomePage
-		 * go through all elements in list, submit something to each field.
-		 * Have what it submits also count up in some way...
-		 */
-		try {
-			List<WebElement> tempElements = aPHomePage.getWebelements(fieldID);
-
-			// Assert.assertTrue("Couldn't find the list of fields using the field id of: "
-			// + fieldID, aPHomePage.can);
-		} catch (AssertionError e) {
-		}
-
-		try {
-			Assert.assertTrue("Couldn't set the field of the practiceFormPage with the field id: " + fieldID,
-					aPHomePage.trySetField(userInput, fieldID));
-		} catch (AssertionError e) {
-			practiceFormPage.takeScreenShot(fieldID);
-			Assert.fail("Test failed due to AssertionError: " + e.getMessage());
-		}
-	}
-
-	@Then("each {string} should contain {string}")
-	public void each_should_contain(String s, String s2) {
-		// Write code here that turns the phrase above into concrete actions
-	}
-
-	@Given("the user clicks on each {string} in the form.")
-	public void the_user_clicks_on_each_in_the_form(String s) {
-		// Write code here that turns the phrase above into concrete actions
-	}
-
-	@Given("the user enters {string} into the {string} field in the form")
-	public void the_user_enters_into_the_field_in_the_form(String userInput, String fieldID) {
-
-	}
-
-	@Then("{string} should contain {string}")
-	public void should_contain(String s, String s2) {
-		// Write code here that turns the phrase above into concrete actions
-	}
-
-	@And("the user enters {string} into the {string} field in the iframe")
-	public void the_user_enters_into_the_field_in_the_iframe(String userInput, String fieldID) {
-
-	}
-
-	@Then("the {string} field in the iframe should contain {string}")
-	public void the_field_in_the_iframe_should_contain(String fieldID, String expectedValue) {
-		try {
-			// TODO: Go back through and check that you use Assert.fail() in catches.
-			Assert.assertTrue(
-					"The name field on the practiceFormPage did not have the expected value. Practice form: "
-							+ practiceFormPage.getFieldValue(fieldID) + ", expected value: " + expectedValue,
-					practiceFormPage.getFieldValue(fieldID).toLowerCase().contains(expectedValue.toLowerCase()));
-		} catch (AssertionError e) {
-			practiceFormPage.takeScreenShot(fieldID);
-			Assert.fail("Test failed due to AssertionError: " + e.getMessage());
-		}
-	}
-
 	@Given("the user selects {string} from the {string}")
 	public void the_user_selects_from_the(String userSelection, String elementID) {
 
 		try {
 			Assert.assertTrue("Couldn't find the element using the id: " + elementID,
-					practiceFormPage.canFindWebelement(elementID));
+					aPHomePage.canFindWebelement(elementID));
 		} catch (AssertionError e) {
 			Assert.fail("Test failed due to AssertionError: " + e.getMessage());
 		}
 
 		// TODO: Need to work out if the general takeScreenShot method works now or not.
 		try {
-			Assert.assertTrue(practiceFormPage.trySelectInDropdown(userSelection, elementID));
+			Assert.assertTrue(aPHomePage.trySelectInDropdown(userSelection, elementID));
 		} catch (AssertionError e) {
 			practiceFormPage.takeScreenShot(elementID);
 			Assert.fail("Test failed due to AssertionError: " + e.getMessage());
@@ -510,7 +428,7 @@ public class InteractingWithOtherElements {
 		// So, create has function that tries to set these then, and that returns true
 		// provided it can find it and turn it into a select.
 		try {
-			Assert.assertTrue(practiceFormPage.trySelectInDropdown(TestConstants.DOBDROPDOWN_ID, year,
+			Assert.assertTrue(aPHomePage.trySelectInDropdown(TestConstants.DOBDROPDOWN_ID, year,
 					Integer.parseInt(month), day));
 		} catch (AssertionError e) {
 			// TODO: handle exception
