+++ src/test/java/resources/features/FillingInFields.feature
@@ -1,4 +1,4 @@
-@skip
+
 Feature: Filling in fields
 
   Scenario: Reading from homepage
