--- src/test/java/resources/features/NextFeatures.feature
@@ -1,7 +1,0 @@
-@skip
-Feature: Next Feature
-
-  Scenario: Change size of area
-    Given the user is on the correct page
-    And the user sees the container for "changing size"
-    Then the user uses the cursor to change the size of the object
